Set Populate start time to time of last entry
It made no sense to me why the populate start time was always set to the
current time, even when there were existing schedule events in the
future. This would cause overwriting of events which I can think of no
useful reason for.

Instead, this gets the last event under the current team by end time,
and then sets the populate start time to that. If there isn't one (e.g.
this is a new schedule), it will still default to now.